requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
aiohttp>=3.9.0
httpx>=0.27.0
orjson>=3.10.0
numpy>=1.26.0
pyyaml>=6.0.1
jinja2>=3.1.0
websockets>=12.0

